kernel: replace unsafe uses of 1L
On some 64bit systems (most notably, Windows), sizeof(long) == 4, and so
expressions like `1L << 60` overflow. Moreover, an enum strictly
speaking is only guaranteed to be able to represent an int, and so using
it to represent 64bit constants is not safe; we hence turn
INT_INTOBJ_MIN and INT_INTOBJ_MAX into global static constants.

+++ src/macfloat.c
@@ -431,7 +431,7 @@ static Obj FuncINTFLOOR_MACFLOAT(Obj self, Obj macfloat)
 #endif
 
 
-  if (fabs(f) < (Double) (1L<<NR_SMALL_INT_BITS))
+  if (fabs(f) < (Double)((Int)1 << NR_SMALL_INT_BITS))
     return INTOBJ_INT((Int)f);
 
   int str_len = (int) (log(fabs(f)) / log(16.0)) + 3;
